Remove uint32_t point parameter from parse declaration
Apologies if I misunderstood something, but it seems that the `parse` declaration in `lexer.h` has a `point` parameter, which is missing from the implementation in `lexer.c`, and doesn't seem to be referenced anywhere. So this PR removes the parameter from `lexer.h`.

The emscripten toolchain I'm using raises this a `conflicting types for 'parse'` error.